Astar maze improvements

+++ src/model/algorithm/H_functions.java
@@ -2,6 +2,6 @@ package model.algorithm;
 
 public interface H_functions {
 
-	double h_function(State current , State goal) ;
+	int h_function(State current , State goal) ;
 
 }
